fix: review fixes, renames
